rootProject 이름 변경
