+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'basic'
+rootProject.name = 'spring-basic'
