--- src/main/java/com/programmers/basic/controller/dto/ListVouchersResponse.java
@@ -1,4 +1,0 @@
-package com.programmers.basic.controller.dto;
-
-public class ListVouchersResponse {
-}
