--- src/main/java/com/programmers/basic/entity/FixedAmountVoucher.java
@@ -1,28 +1,0 @@
-package com.programmers.basic.entity;
-
-import java.util.UUID;
-
-public class FixedAmountVoucher implements Voucher {
-
-	private final UUID voucherId;
-	private final long amount;
-
-	public FixedAmountVoucher(UUID voucherId, long amount) {
-		this.voucherId = voucherId;
-		this.amount = amount;
-	}
-
-	@Override
-	public UUID getVoucherId() {
-		return voucherId;
-	}
-
-	@Override
-	public String toString() {
-		return "FixedAmountVoucher{" +
-			"voucherId=" + voucherId +
-			", amount=" + amount +
-			'}';
-	}
-
-}
