--- src/main/java/com/programmers/basic/entity/Voucher.java
@@ -1,7 +1,0 @@
-package com.programmers.basic.entity;
-
-import java.util.UUID;
-
-public interface Voucher {
-	UUID getVoucherId();
-}
