+++ src/main/java/com/programmers/springbasic/VoucherApplication.java
@@ -1,13 +1,13 @@
-package com.programmers.basic;
+package com.programmers.springbasic;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class SpringbootBasicApplication {
+public class VoucherApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(SpringbootBasicApplication.class, args);
+		SpringApplication.run(VoucherApplication.class, args);
 	}
 
 }
