+++ src/main/java/com/programmers/springbasic/configuration/CommandExecutorConfig.java
@@ -1,10 +1,10 @@
-package com.programmers.basic.configuration;
+package com.programmers.springbasic.configuration;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.boot.CommandLineRunner;
 
-import com.programmers.basic.command.CommandExecutor;
+import com.programmers.springbasic.command.CommandExecutor;
 
 @Configuration
 public class CommandExecutorConfig {
