+++ src/main/java/com/programmers/springbasic/controller/VoucherController.java
@@ -1,0 +1,34 @@
+package com.programmers.springbasic.controller;
+
+import java.util.List;
+import java.util.UUID;
+
+import org.springframework.stereotype.Controller;
+
+import com.programmers.springbasic.controller.dto.CreateFixedAmountVoucherRequest;
+import com.programmers.springbasic.controller.dto.CreatePercentDiscountVoucherRequest;
+import com.programmers.springbasic.controller.dto.ListVouchersResponse;
+import com.programmers.springbasic.service.VoucherService;
+
+@Controller
+public class VoucherController {
+
+	private final VoucherService voucherService;
+
+	public VoucherController(VoucherService voucherService) {
+		this.voucherService = voucherService;
+	}
+
+	public UUID createPercentDiscountVoucher(CreatePercentDiscountVoucherRequest request) {
+		return voucherService.createPercentDiscountVoucher(request);
+	}
+
+	public UUID createFixedAmountVoucher(CreateFixedAmountVoucherRequest request) {
+		return voucherService.createFixedAmountVoucher(request);
+	}
+
+	public List<ListVouchersResponse> listVoucher() {
+		return voucherService.listVoucher();
+	}
+
+}
