+++ src/main/java/com/programmers/springbasic/controller/dto/CreatePercentDiscountVoucherRequest.java
@@ -1,9 +1,8 @@
-package com.programmers.basic.controller.dto;
+package com.programmers.springbasic.controller.dto;
 
 import java.util.UUID;
 
-import com.programmers.basic.entity.FixedAmountVoucher;
-import com.programmers.basic.entity.PercentDiscountVoucher;
+import com.programmers.springbasic.entity.PercentDiscountVoucher;
 
 public record CreatePercentDiscountVoucherRequest(long percent) {
 	public PercentDiscountVoucher toEntity(UUID voucherId) {
