+++ src/main/java/com/programmers/springbasic/controller/dto/ListVouchersResponse.java
@@ -1,0 +1,16 @@
+package com.programmers.springbasic.controller.dto;
+
+import java.util.UUID;
+
+import com.programmers.springbasic.entity.Voucher;
+import com.programmers.springbasic.entity.VoucherType;
+
+public record ListVouchersResponse(
+	UUID voucherId,
+	VoucherType voucherType,
+	long amount,
+	long percent) {
+	// public static ListVouchersResponse from(Voucher voucher) {
+	//
+	// }
+}
