+++ src/main/resources/application.properties
@@ -1,1 +1,2 @@
 
+spring.main.web-application-type=none
