+++ src/test/java/com/programmers/springbasic/SpringbootBasicApplicationTests.java
@@ -1,4 +1,4 @@
-package com.programmers.basic;
+package com.programmers.springbasic;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
